Add files via upload
Sorry this came out so late, I haven't had much time to work on the presentation, but I made sure to include things that I felt were important to discuss. It's a little messy now, but I'll probably continue working on it until we have to present. Please add/remove things to/from the presentation if it needs to be done.